設定永続化実装

+++ Assets/_Project/Scripts/Config/MatchSetup/MatchSetupNetworkCodec.cs
@@ -41,7 +41,7 @@ namespace DiceGame.Config
                 return false;
             }
 
-            if (!System.Enum.IsDefined(typeof(GameMode), payload.GameMode)) {
+            if (!System.Enum.IsDefined(typeof(GameMode), (int)payload.GameMode)) {
                 errorMessage = "MatchSetupNetworkCodec: Unknown game mode.";
                 return false;
             }
